Add horizontal scroll

+++ components/MyProjectsSection/MyProjectsSection.tsx
@@ -13,43 +13,65 @@ const MyProjectsSection = () => {
   gsap.registerPlugin(ScrollTrigger);
 
   useEffect(() => {
-    const pin = gsap.fromTo(
-      sectionRef.current,
-      {
-        translateX: 0,
-      },
-      {
-        translateX: "-100%",
-        ease: "none",
-        duration: 1,
-        scrollTrigger: {
-          trigger: triggerRef.current,
-          scrub: 0.6,
-          pin: true,
+    let mm = gsap.matchMedia();
+    let pin: any = null;
+
+    mm.add("(min-width: 768px)", () => {
+      pin = gsap.fromTo(
+        sectionRef.current,
+        {
+          translateX: 0,
         },
-      }
-    );
+        {
+          translateX: "-200vw",
+          ease: "none",
+          duration: 3,
+          scrollTrigger: {
+            trigger: triggerRef.current,
+            scrub: 0.6,
+            pin: true,
+          },
+        }
+      );
+    });
+
+    mm.add("(max-width: 768px)", () => {
+      pin = gsap.fromTo(
+        sectionRef.current,
+        {
+          translateY: 0,
+        },
+        {
+          translateY: "-200vh",
+          ease: "none",
+          duration: 2,
+          scrollTrigger: {
+            trigger: triggerRef.current,
+            scrub: 0.0,
+            pin: true,
+          },
+        }
+      );
+    });
 
     return () => {
-      pin.kill();
+      pin && pin.kill();
     };
   }, []);
 
   return (
-    <div>
+    <div className='overflow-hidden h-[200vh]' ref={triggerRef}>
       <div className='sticky top-0'>
         <TextSlider text='MY WORK' />
         <MyWorkGradientBackground />
       </div>
-      <section className='overflow-x-hidden h-screen' ref={triggerRef}>
-        <div className='flex items-center relative' ref={sectionRef}>
-          {projectPaneTexts.map((projectPaneText) => (
-            <div className='w-screen flex-none'>
-              <ProjectPane {...projectPaneText} />
-            </div>
-          ))}
-        </div>
-      </section>
+      <div className='flex flex-col md:flex-row ' ref={sectionRef}>
+        {projectPaneTexts.map((projectPaneText) => (
+          <div className='w-screen flex-none'>
+            <ProjectPane {...projectPaneText} />
+          </div>
+        ))}
+      </div>
     </div>
   );
 };
